refactor: move Malgo/{UTerm.hs, Unify.hs} to Malgo/Infer/
