+++ malgo/malgo.cabal
@@ -29,6 +29,8 @@ library
       Malgo.Driver
       Malgo.Infer.Pass
       Malgo.Infer.TcEnv
+      Malgo.Infer.Unify
+      Malgo.Infer.UTerm
       Malgo.Interface
       Malgo.Parser
       Malgo.Prelude
@@ -41,8 +43,6 @@ library
       Malgo.Syntax.Extension
       Malgo.TypeRep.Static
       Malgo.TypeRep.UTerm
-      Malgo.Unify
-      Malgo.UTerm
   other-modules:
       Paths_malgo
   hs-source-dirs:
